refactor(다리 건너기): 클래스별로 패키지 분류
