+++ src/main/java/bridge/Application.java
@@ -1,5 +1,10 @@
 package bridge;
 
+import bridge.controller.BridgeGameController;
+import bridge.domain.BridgeValidator;
+import bridge.view.InputView;
+import bridge.view.OutputView;
+
 public class Application {
 
     public static void main(String[] args) {
