+++ src/main/java/bridge/domain/BridgeMaker.java
@@ -1,5 +1,6 @@
-package bridge;
+package bridge.domain;
 
+import bridge.BridgeNumberGenerator;
 import java.util.ArrayList;
 import java.util.List;
