+++ src/main/java/bridge/domain/BridgeValidator.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.domain;
 
 public class BridgeValidator {
