+++ src/main/java/bridge/domain/GameStatus.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.domain;
 
 public enum GameStatus {
     SUCCESS("성공"), FAIL("실패"),
