+++ src/test/java/bridge/BridgeGameTest.java
@@ -3,6 +3,7 @@ package bridge;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import bridge.domain.BridgeGame;
 import bridge.exception.CannotCrossBridgeException;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
