+++ src/test/java/bridge/BridgeValidatorTest.java
@@ -2,6 +2,7 @@ package bridge;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import bridge.domain.BridgeValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
